+++ HISTORY.md
@@ -48,6 +48,9 @@
 * **Translation Updates:**
     * German
 
+### 2022-05-017
+* blocks: added experimental private isChangeableTo(type) method
+
 ### 2022-05-06
 * threads: include currently dragged sprites in the MY OTHER SPRITES/CLONES lists